# [sparse] use _sparse_coo_tensor_unsafe in coalesce for speedup